+++ base/src/config/api.rs
@@ -7,11 +7,19 @@ lazy_static! {
 }
 
 /// Web server configuration options.
-#[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Serialize, Deserialize)]
+#[derive(Clone, Eq, PartialEq, Hash, Debug, Serialize, Deserialize)]
 pub struct APIConfig {
     /// Local addess to bind the API server to.
     #[serde(default = "APIConfig::default_bind")]
     pub bind: String,
+
+    /// The number of request handling threads.
+    #[serde(default)]
+    pub threads_count: Option<usize>,
+
+    /// API server timeouts.
+    #[serde(default)]
+    pub timeouts: Timeouts,
 
     /// Enable/disable entire API trees.
     #[serde(default)]
@@ -22,6 +30,8 @@ impl Default for APIConfig {
     fn default() -> Self {
         APIConfig {
             bind: Self::default_bind(),
+            threads_count: None,
+            timeouts: Timeouts::default(),
             trees: APITrees::default(),
         }
     }
@@ -50,7 +60,7 @@ impl APIConfig {
     pub fn set_default_bind(bind: String) {
         let mut default = DEFAULT_BIND.write().unwrap();
         if default.is_some() {
-            panic!("Cannot override the default api.bind option more than once");
+            panic!("cannot override the default api.bind option more than once");
         }
         *default = Some(bind);
     }
@@ -92,5 +102,45 @@ impl From<APITrees> for HashMap<&'static str, bool> {
         flags.insert("introspect", trees.introspect);
         flags.insert("unstable", trees.unstable);
         flags
+    }
+}
+
+/// API server timeouts.
+#[derive(Clone, Eq, PartialEq, Hash, Debug, Serialize, Deserialize)]
+pub struct Timeouts {
+    /// Control the timeout, in seconds, for keep alive connections.
+    #[serde(default = "Timeouts::default_keep_alive")]
+    pub keep_alive: Option<u64>,
+
+    /// Control the timeout, in seconds, for reads on existing connections.
+    #[serde(default = "Timeouts::default_read")]
+    pub read: Option<u64>,
+
+    /// Control the timeout, in seconds, for writes on existing connections.
+    #[serde(default = "Timeouts::default_write")]
+    pub write: Option<u64>,
+}
+
+impl Default for Timeouts {
+    fn default() -> Timeouts {
+        Timeouts {
+            keep_alive: Self::default_keep_alive(),
+            read: Self::default_read(),
+            write: Self::default_write(),
+        }
+    }
+}
+
+impl Timeouts {
+    fn default_keep_alive() -> Option<u64> {
+        Some(5)
+    }
+
+    fn default_read() -> Option<u64> {
+        Some(5)
+    }
+
+    fn default_write() -> Option<u64> {
+        Some(1)
     }
 }