+++ zookeeper/src/lib.rs
@@ -17,18 +17,7 @@ extern crate replicante_agent;
 extern crate replicante_agent_models;
 extern crate replicante_util_tracing;
 
-use std::time::Duration;
-
-use clap::App;
-use clap::Arg;
-use failure::ResultExt;
-
-use replicante_agent::AgentContext;
-use replicante_agent::AgentRunner;
 use replicante_agent::Result;
-
-use replicante_util_tracing::tracer;
-use replicante_util_tracing::TracerExtra;
 
 mod agent;
 mod config;
@@ -38,7 +27,6 @@ mod zk4lw;
 
 use agent::ZookeeperAgent;
 use config::Config;
-use error::ErrorKind;
 
 lazy_static! {
     /// Version string.
@@ -53,50 +41,26 @@ lazy_static! {
 const DEFAULT_CONFIG_FILE: &str = "agent-zookeeper.yaml";
 
 /// Configure and start the agent.
-pub fn run() -> Result<()> {
+pub fn run() -> Result<bool> {
     // Command line parsing.
-    let cli_args = App::new("Zookeeper Replicante Agent")
-        .version(VERSION.as_ref())
-        .about(env!("CARGO_PKG_DESCRIPTION"))
-        .arg(
-            Arg::with_name("config")
-                .short("c")
-                .long("config")
-                .value_name("FILE")
-                .default_value(DEFAULT_CONFIG_FILE)
-                .help("Specifies the configuration file to use")
-                .takes_value(true),
-        )
-        .get_matches();
+    let cli_args = ::replicante_agent::process::clap(
+        "Zookeeper Replicante Agent",
+        VERSION.as_ref(),
+        env!("CARGO_PKG_DESCRIPTION"),
+        DEFAULT_CONFIG_FILE,
+    )
+    .get_matches();
 
-    // Load configuration (default file is allowed to be missing).
+    // Load configuration.
     Config::override_defaults();
     let config_location = cli_args.value_of("config").unwrap();
     let config = Config::from_file(config_location)?;
     let config = config.transform();
 
-    // Configure the logger (from the agent context).
-    let agent_config = config.agent.clone();
-    let (logger, _scope_guard) = AgentContext::logger(&agent_config);
-
-    // Setup and run the tracer.
-    let (tracer, mut extra) = tracer(config.agent.tracing.clone(), logger.clone())
-        .with_context(|_| ErrorKind::Initialisation("tracer configuration failed".into()))?;
-    if let TracerExtra::ReporterThread(ref mut reporter) = extra {
-        reporter.stop_delay(Duration::from_secs(2));
-    }
-
-    // Setup the agent context.
-    let agent_context = AgentContext::new(agent_config, logger, tracer);
-    AgentRunner::register_metrics(&agent_context.logger, &agent_context.metrics);
-    metrics::register_metrics(&agent_context.logger, &agent_context.metrics);
-
-    // Setup and run the agent.
-    let agent = ZookeeperAgent::new(config, agent_context.clone());
-    let runner = AgentRunner::new(agent, agent_context);
-    runner.run();
-
-    // Cleanup tracer and exit.
-    drop(extra);
-    Ok(())
+    // Run the agent using the provided default helper.
+    ::replicante_agent::process::run(config.agent.clone(), |context, _, _| {
+        metrics::register_metrics(context);
+        let agent = ZookeeperAgent::new(config, context.clone());
+        Ok(agent)
+    })
 }